output edit

+++ src/main/scala/Hello.scala
@@ -28,13 +28,13 @@ object Hello  {
   def max2(a: (Long, Long), b: (Long, Long)): (Long, Long) = if (a._2 > b._2) a else b
 
 
-  def percolation(N: Long, T: Long, nTries: Int, sc: SparkContext, edgeRate: Int, rndm: Random, randomEdges: (scala.util.Random, Int, Int) => List[Int], writer: PrintWriter): Unit = {
+  def percolation(N: Int, T: Int, nTries: Int, sc: SparkContext, edgeRate: Int, rndm: Random, randomEdges: (scala.util.Random, Int, Int) => List[Int], writer: PrintWriter): Unit = {
 
     /*
     initialize N vertices; each vertex has 2 attributes: Id (Long), name(String)
     */
     val verticesTemp = ListBuffer[(Long, String)]()
-    for (i <- 0 until N.toInt) {
+    for (i <- 0 until N) {
       verticesTemp += ((i, i.toString + "th"))
     }
     val vRDD = sc.parallelize(verticesTemp.toList) // form RDD of the vertices
@@ -52,13 +52,13 @@ object Hello  {
      */
     writer.write("t|maxTris|avgDegree|maxCCsize\n")
 
-    for (t <- 0 until T.toInt ) {
+    for (t <- 0 until T ) {
       /*
       To add edgeRate number of new edges during each t, make a bucket to hold new edges that are randomly generated.
       Initialize the bucket with one new randomly edge. bucket size allows to set the rate of increase in size of network
       time per step
        */
-      val newEdges = randomEdges(rndm,N.toInt,nTries)
+      val newEdges = randomEdges(rndm,N,nTries)
       var v1 = newEdges(0)
       var v2 = newEdges(1)
       val nextEdge = ArrayBuffer(Edge(v1.toLong, v2.toLong, 1), Edge(v2.toLong, v1.toLong, 1))
@@ -67,7 +67,7 @@ object Hello  {
       val edgeBucket = nextEdge
       // Add remaining edgeRate-1 randomly generated edges to the bucket
       for (n <- 1 to edgeRate) {
-        val currEdges = randomEdges(rndm, N.toInt, nTries)
+        val currEdges = randomEdges(rndm, N, nTries)
         v1 = currEdges(0)
         v2 = currEdges(1)
         edgeBucket +=(Edge(v1.toLong, v2.toLong, 1), Edge(v2.toLong, v1.toLong, 1))
@@ -106,8 +106,8 @@ object Hello  {
     val conf = new SparkConf().setAppName("hello-spark").setMaster("local[*]")
     val sc = new SparkContext(conf)
 
-    val T: Long = 100 // the number of events to run; the evolution time of the dynamics
-    val N: Long = 100 // the number of vertices
+    val T: Int = 100 // the number of events to run; the evolution time of the dynamics
+    val N: Int = 100 // the number of vertices
     val edgeRate: Int = 5 // the number of edges to add during each time step
     val nTries: Int = 5   // the number of times to avoid self-links; to avoid infinite while loop
     val rndm = scala.util.Random
